+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Refine include guard macro name for `analysis.hpp`
 - Add `ExpWeightedIirFilter` class
+- Add `printRealMat()` and `printRealVec()` functions
 
 ## [0.1.0] - 2021/7/22